file reorg to fix compilation time problem
